super cleaning...

+++ Sources/hoa.decoder_tilde.cpp
@@ -34,8 +34,8 @@ static t_eclass *hoa_decoder_3d_class;
 
 static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
 {
-    int	order = 1;
-    int arg   = 4;
+    ulong order = 1;
+    ulong arg   = 4;
     t_symbol* mode = gensym("regular");
     t_hoa_decoder *x = (t_hoa_decoder *)eobj_new(hoa_decoder_class);
     t_binbuf *d      = binbuf_via_atoms(argc,argv);
@@ -44,7 +44,7 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
 	{
         if(argc && argv && atom_gettype(argv) == A_LONG)
         {
-            order = pd_clip_minmax(atom_getlong(argv), 1, 63);
+            order = ulong(pd_clip_minmax(atom_getlong(argv), 1, 63));
         }
         if(argc > 1 && argv+1 && atom_gettype(argv+1) == A_SYM)
             mode = atom_getsym(argv+1);
@@ -54,7 +54,7 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
             arg = order * 2 + 1;
             if(argc > 2 && argv+2 && atom_gettype(argv+2) == A_LONG)
             {
-                arg = pd_clip_min(atom_getlong(argv+2), 1);
+                arg = ulong(pd_clip_min(atom_getlong(argv+2), 1));
             }
             x->f_decoder = new Decoder<Hoa2d, t_sample>::Irregular(order, arg);
             x->f_mode = mode;
@@ -75,7 +75,7 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
             x->f_mode = gensym("regular");
         }
 
-        eobj_dspsetup(x, x->f_decoder->getNumberOfHarmonics(), x->f_decoder->getNumberOfPlanewaves());
+        eobj_dspsetup(x, long(x->f_decoder->getNumberOfHarmonics()), long(x->f_decoder->getNumberOfPlanewaves()));
         x->f_ins = new t_sample[x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE];
         x->f_outs= new t_sample[x->f_decoder->getNumberOfPlanewaves() * HOA_MAXBLKSIZE];
         memset(x->f_ins, 0, x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE * sizeof(t_sample));
@@ -92,7 +92,7 @@ static void hoa_decoder_perform64_regular(t_hoa_decoder *x, t_object *dsp64, t_s
 {
     for(long i = 0; i < numins; i++)
     {
-        Signal<t_sample>::copy(sampleframes, ins[i], 1, x->f_ins+i, numins);
+        Signal<t_sample>::copy(ulong(sampleframes), ins[i], 1, x->f_ins+i, ulong(numins));
     }
 	for(long i = 0; i < sampleframes; i++)
     {
@@ -100,7 +100,7 @@ static void hoa_decoder_perform64_regular(t_hoa_decoder *x, t_object *dsp64, t_s
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_outs+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_outs+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -108,7 +108,7 @@ static void hoa_decoder_perform64_irregular(t_hoa_decoder *x, t_object *dsp64, t
 {
     for(long i = 0; i < numins; i++)
     {
-        Signal<t_sample>::copy(sampleframes, ins[i], 1, x->f_ins+i, numins);
+        Signal<t_sample>::copy(ulong(sampleframes), ins[i], 1, x->f_ins+i, ulong(numins));
     }
     for(long i = 0; i < sampleframes; i++)
     {
@@ -116,7 +116,7 @@ static void hoa_decoder_perform64_irregular(t_hoa_decoder *x, t_object *dsp64, t
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_outs+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_outs+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -127,7 +127,7 @@ static void hoa_decoder_perform64_binaural(t_hoa_decoder *x, t_object *dsp64, t_
 
 static void hoa_decoder_dsp(t_hoa_decoder *x, t_object *dsp64, short *count, double samplerate, long maxvectorsize, long flags)
 {
-    x->f_decoder->computeRendering(maxvectorsize);
+    x->f_decoder->computeRendering(ulong(maxvectorsize));
     if(x->f_mode == gensym("binaural"))
     {
         object_method(dsp64, gensym("dsp_add64"), x, (method)hoa_decoder_perform64_binaural, 0, NULL);
@@ -150,7 +150,7 @@ static t_pd_err hoa_decoder_angles_set(t_hoa_decoder *x, void *attr, long argc, 
         for(long i = 0; i < argc && i < (long)x->f_decoder->getNumberOfPlanewaves(); i++)
         {
             if(atom_gettype(argv+i) == A_FLOAT)
-                x->f_decoder->setPlanewaveAzimuth(i, atom_getfloat(argv+i) / 360. * HOA_2PI);
+                x->f_decoder->setPlanewaveAzimuth(ulong(i), atom_getfloat(argv+i) / 360. * HOA_2PI);
         }
         canvas_resume_dsp(dspState);
     }
@@ -160,13 +160,13 @@ static t_pd_err hoa_decoder_angles_set(t_hoa_decoder *x, void *attr, long argc, 
 
 static t_pd_err hoa_decoder_angles_get(t_hoa_decoder *x, void *attr, long* argc, t_atom **argv)
 {
-    *argc = x->f_decoder->getNumberOfPlanewaves();
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argc = long(x->f_decoder->getNumberOfPlanewaves());
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         for(int i = 0; i < *argc; i++)
         {
-             atom_setfloat(*argv+i, x->f_decoder->getPlanewaveAzimuth(i) * 360. / HOA_2PI);
+             atom_setfloat(*argv+i, x->f_decoder->getPlanewaveAzimuth(ulong(i)) * 360. / HOA_2PI);
         }
     }
     else
@@ -191,7 +191,7 @@ static t_pd_err hoa_decoder_offset_set(t_hoa_decoder *x, void *attr, long argc, 
 static t_pd_err hoa_decoder_offset_get(t_hoa_decoder *x, void *attr, long* argc, t_atom **argv)
 {
     *argc = 1;
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         atom_setfloat(*argv, x->f_decoder->getPlanewavesRotationZ() * 360. / HOA_2PI);
@@ -221,7 +221,7 @@ static t_pd_err hoa_decoder_crop_set(t_hoa_decoder *x, void *attr, long argc, t_
 static t_pd_err hoa_decoder_crop_get(t_hoa_decoder *x, void *attr, long* argc, t_atom **argv)
 {
     *argc = 1;
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         if(x->f_mode == gensym("binaural"))
@@ -284,9 +284,9 @@ extern "C" void setup_hoa0x2e2d0x2edecoder_tilde(void)
 
 static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
 {
-    int	order = 1;
+    ulong order = 1;
     t_symbol* mode = gensym("regular");
-    int number_of_channels = 4;
+    ulong number_of_channels = 4;
     t_hoa_decoder_3d *x = (t_hoa_decoder_3d *)eobj_new(hoa_decoder_3d_class);
     t_binbuf *d         = binbuf_via_atoms(argc,argv);
 
@@ -294,13 +294,13 @@ static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
     {
         if(argc && argv && atom_gettype(argv) == A_LONG)
         {
-            order = pd_clip_minmax(atom_getlong(argv), 1, 10);
+            order = ulong(pd_clip_minmax(atom_getlong(argv), 1, 10));
             number_of_channels = (order+1)*(order+1);
         }
         if(argc > 1 && argv+1 && atom_gettype(argv+1) == A_SYM)
             mode = atom_getsym(argv+1);
         if(argc > 2 && argv+2 && atom_gettype(argv+2) == A_LONG)
-            number_of_channels = pd_clip_min(atom_getlong(argv+2), 1);
+            number_of_channels = ulong(pd_clip_min(atom_getlong(argv+2), 1));
 
         if(mode == gensym("irregular"))
         {
@@ -318,7 +318,7 @@ static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
             x->f_mode = mode;
         }
 
-        eobj_dspsetup(x, x->f_decoder->getNumberOfHarmonics(), x->f_decoder->getNumberOfPlanewaves());
+        eobj_dspsetup(x, long(x->f_decoder->getNumberOfHarmonics()), long(x->f_decoder->getNumberOfPlanewaves()));
         x->f_ins = new t_sample[x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE];
         x->f_outs= new t_sample[x->f_decoder->getNumberOfPlanewaves() * HOA_MAXBLKSIZE];
         memset(x->f_ins, 0, x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE * sizeof(t_sample));
@@ -335,7 +335,7 @@ static void hoa_decoder_3d_perform64_regular(t_hoa_decoder_3d *x, t_object *dsp6
 {
     for(long i = 0; i < numins; i++)
     {
-        Signal<t_sample>::copy(sampleframes, ins[i], 1, x->f_ins+i, numins);
+        Signal<t_sample>::copy(ulong(sampleframes), ins[i], 1, x->f_ins+i, ulong(numins));
     }
     for(long i = 0; i < sampleframes; i++)
     {
@@ -343,7 +343,7 @@ static void hoa_decoder_3d_perform64_regular(t_hoa_decoder_3d *x, t_object *dsp6
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_outs+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_outs+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -354,7 +354,7 @@ static void hoa_decoder_3d_perform64_binaural(t_hoa_decoder_3d *x, t_object *dsp
 
 static void hoa_decoder_3d_dsp(t_hoa_decoder *x, t_object *dsp64, short *count, double samplerate, long maxvectorsize, long flags)
 {
-    x->f_decoder->computeRendering(maxvectorsize);
+    x->f_decoder->computeRendering(ulong(maxvectorsize));
     if(x->f_mode == gensym("binaural"))
     {
         object_method(dsp64, gensym("dsp_add64"), x, (method)hoa_decoder_3d_perform64_binaural, 0, NULL);
@@ -376,12 +376,12 @@ static t_pd_err hoa_decoder_3d_angles_set(t_hoa_decoder_3d *x, void *attr, long 
             {
                 if(j%2)
                 {
-                    x->f_decoder->setPlanewaveElevation(i, atom_getfloat(argv+j) / 360. * HOA_2PI);
+                    x->f_decoder->setPlanewaveElevation(ulong(i), atom_getfloat(argv+j) / 360. * HOA_2PI);
                     i++;
                 }
                 else
                 {
-                    x->f_decoder->setPlanewaveAzimuth(i, atom_getfloat(argv+j) / 360. * HOA_2PI);
+                    x->f_decoder->setPlanewaveAzimuth(ulong(i), atom_getfloat(argv+j) / 360. * HOA_2PI);
                 }
 
             }
@@ -394,8 +394,8 @@ static t_pd_err hoa_decoder_3d_angles_set(t_hoa_decoder_3d *x, void *attr, long 
 
 static t_pd_err hoa_decoder_3d_angles_get(t_hoa_decoder_3d *x, void *attr, long *argc, t_atom **argv)
 {
-    *argc = x->f_decoder->getNumberOfPlanewaves() * 2;
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argc = long(x->f_decoder->getNumberOfPlanewaves() * 2);
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         for(ulong i = 0; i < x->f_decoder->getNumberOfPlanewaves(); i++)
@@ -439,7 +439,7 @@ static t_pd_err hoa_decoder_3d_offset_set(t_hoa_decoder_3d *x, void *attr, long 
 static t_pd_err hoa_decoder_3d_offset_get(t_hoa_decoder_3d *x, void *attr, long *argc, t_atom **argv)
 {
     *argc = 3;
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         atom_setfloat(*argv, x->f_decoder->getPlanewavesRotationX() / 360. * HOA_2PI);
@@ -471,7 +471,7 @@ static t_pd_err hoa_decoder_3d_crop_set(t_hoa_decoder_3d *x, void *attr, long ar
 static t_pd_err hoa_decoder_3d_crop_get(t_hoa_decoder_3d *x, void *attr, long* argc, t_atom **argv)
 {
     *argc = 1;
-    *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
+    *argv = (t_atom *)malloc(size_t(*argc) * sizeof(t_atom));
     if(*argc && *argv)
     {
         if(x->f_mode == gensym("binaural"))
