+++ Sources/hoa.encoder_tilde.cpp
@@ -28,15 +28,15 @@ static t_eclass *hoa_encoder_3d_class;
 
 static void *hoa_encoder_new(t_symbol *s, long argc, t_atom *argv)
 {
-	int	order = 1;
+	ulong order = 1;
     t_hoa_encoder *x = (t_hoa_encoder *)eobj_new(hoa_encoder_class);
 	if (x)
 	{
         if(atom_gettype(argv) == A_LONG)
-            order = pd_clip_minmax(atom_getlong(argv), 1, 63);
+            order = ulong(pd_clip_minmax(atom_getlong(argv), 1, 63));
         
         x->f_encoder = new Encoder<Hoa2d, t_sample>::Basic(order);
-        eobj_dspsetup(x, 2, x->f_encoder->getNumberOfHarmonics());
+        eobj_dspsetup(x, 2, long(x->f_encoder->getNumberOfHarmonics()));
         
         x->f_signals =  new t_sample[x->f_encoder->getNumberOfHarmonics() * 8192];
         return x;
@@ -58,7 +58,7 @@ static void hoa_encoder_perform(t_hoa_encoder *x, t_object *dsp, t_sample **ins,
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -70,7 +70,7 @@ static void hoa_encoder_perform_offset(t_hoa_encoder *x, t_object *dsp, t_sample
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -107,7 +107,7 @@ extern "C" void setup_hoa0x2e2d0x2eencoder_tilde(void)
 
 static void *hoa_encoder_3d_new(t_symbol *s, long argc, t_atom *argv)
 {
-    int	order = 1;
+    ulong order = 1;
     t_hoa_encoder_3d *x = (t_hoa_encoder_3d *)eobj_new(hoa_encoder_3d_class);
     if(x)
     {
@@ -115,7 +115,7 @@ static void *hoa_encoder_3d_new(t_symbol *s, long argc, t_atom *argv)
             order = pd_clip_minmax(atom_getlong(argv), 1, 10);
         
         x->f_encoder = new Encoder<Hoa3d, t_sample>::Basic(order);
-        eobj_dspsetup(x, 3, x->f_encoder->getNumberOfHarmonics());
+        eobj_dspsetup(x, 3, long(x->f_encoder->getNumberOfHarmonics()));
         
         x->f_signals =  new t_sample[x->f_encoder->getNumberOfHarmonics() * 8192];
     }
@@ -140,7 +140,7 @@ static void hoa_encoder_3d_perform(t_hoa_encoder_3d *x, t_object *dsp, float **i
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -153,7 +153,7 @@ static void hoa_encoder_3d_perform_azimuth(t_hoa_encoder_3d *x, t_object *dsp, t
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -166,7 +166,7 @@ static void hoa_encoder_3d_perform_elevation(t_hoa_encoder_3d *x, t_object *dsp,
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
 
@@ -178,7 +178,7 @@ static void hoa_encoder_3d_perform_offset(t_hoa_encoder_3d *x, t_object *dsp, t_
     }
     for(long i = 0; i < numouts; i++)
     {
-        Signal<t_sample>::copy(sampleframes, x->f_signals+i, numouts, outs[i], 1);
+        Signal<t_sample>::copy(ulong(sampleframes), x->f_signals+i, ulong(numouts), outs[i], 1);
     }
 }
